final upd+exception handler

+++ src/main/java/by/task/telegrambot/model/City.java
@@ -54,7 +54,7 @@ public class City {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
